% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateAlgorithm.R
\name{CalculateAlgorithm}
\alias{CalculateAlgorithm}
\title{Adds two raster lists together}
\usage{
CalculateAlgorithm(BinaryNDVI_List, BinaryRGBSum_List)
}
\arguments{
\item{BinaryNDVI_List}{a list of BinaryNDVI rasters}

\item{BinaryRGBSum_List}{a list of BinaryRGBSum rasters}
}
\value{
returns a list of rasters, each representing an individual sum betweent BinaryNDVI and BinaryRGBSum, with values 0 = no hit on either, 1 = hit on one, 2 = hit on both;
}
\description{
Adds two raster lists together
}
