% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateNDVI_RGB_Combo.R
\name{CalculateCombo}
\alias{CalculateCombo}
\title{Adds two raster lists together, first element with first element and so on, and prints time spent;}
\usage{
CalculateCombo(BinaryNDVI_List, BinaryRGBSum_List)
}
\arguments{
\item{BinaryNDVI_List}{a list of rasters with binary NDVIs, produced earlier based on some Threshold (presumably 12 elements, representing 1 agricultural year);}

\item{BinaryRGBSum_List}{a list of binary rasters with the sum of the RGB bands, produced earlier based on some Threshold (presumably 12 elements, representing 1 agricultural year);}
}
\value{
a list of the sum between the two arguments. Its original use is to return a raster with 3 values, 0 = that pixel hit none of the thresholds, 1 = pixel hit either the NDVI threshold or the RGBSum threshold and 2 = pixel hit both  thresholds;
}
\description{
Adds two raster lists together, first element with first element and so on, and prints time spent;
}
