% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateRGBSum.R
\name{CalculateRGBSum}
\alias{CalculateRGBSum}
\title{Calculates the sum of the first 3 bands of a raster, for each raster in a list and prints time spent;}
\usage{
CalculateRGBSum(BrickList)
}
\arguments{
\item{BrickList}{requires a list of Bricks, R objects that usually store rasters with multiple bands. Should work with a Stack list as well, but untested;}
}
\value{
a list of rasters, with each raster representing the sum of the first 3 bands contained in the corresponding element of the list provided in the argument;
}
\description{
Calculates the sum of the first 3 bands of a raster, for each raster in a list and prints time spent;
}
\examples{
Data is too large, check Vignette.
}
