% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckPackages.R
\name{CheckPackages}
\alias{CheckPackages}
\title{Checks if required packages are installed and loaded and if they're not, installs them and loads them. This is usually not needed as R should install any dependencies; Prints time spent;}
\usage{
CheckPackages()
}
\value{
has no return. Meant to called without arguments;
}
\description{
Checks if required packages are installed and loaded and if they're not, installs them and loads them. This is usually not needed as R should install any dependencies; Prints time spent;
}
\examples{
CheckPackages()
}
