% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_Albufeira_AllYears_List.R
\name{Create_Albufeiras_AllYears_List}
\alias{Create_Albufeiras_AllYears_List}
\title{Organizes the argument into a list of lists. The first element is a list of dataframes that represent all the years of a particular ZOI;}
\usage{
Create_Albufeiras_AllYears_List(Albufeira_Dataframe_List)
}
\arguments{
\item{Albufeira_Dataframe_List}{a list of all the years of imagery, organized by dataframes of ZOI, of the product, usually of Binary_NDVI or Combo; The first element of that list should be a list of all imagery for Year1,
organized by ZOI. So, list first element = ZOI_Dataframe_List_Year1, list second element = ZOI_Dataframe_List_Year2}
}
\value{
returns a list of Dataframe lists, where the first element is a list representing all the dataframes for all the years of the first ZOI;
}
\description{
Organizes the argument into a list of lists. The first element is a list of dataframes that represent all the years of a particular ZOI;
}
\examples{
Data too large, check Vignette.
}
