% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateIndividualFolders_1_2_3.R
\name{CreateIndividualFolders2}
\alias{CreateIndividualFolders2}
\title{Creates folders for each individual ZOI, with threshold imbued in the folder name;}
\usage{
CreateIndividualFolders2(FolderDestinationPath, ID_list, Threshold_List)
}
\arguments{
\item{FolderDestinationPath}{a destination path to the folders being created, as a string;}

\item{ID_list}{a list of IDs representing the amount of unique shapes inside the original shapefile, usually created asa global variable using CropShapefile_ExportToPath;}

\item{Threshold_List}{a list of thresholds, usually just one, representing the threshold for either the NDVI or RGBSum, in order to create properly named folders;}
}
\value{
returns nothing, used to create folders;
}
\description{
Creates folders for each individual ZOI, with threshold imbued in the folder name;
}
