% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateOutputFoldersALL.R
\name{CreateOutputFoldersALL}
\alias{CreateOutputFoldersALL}
\title{Creates some of the folders required to organize outputs in the a certain provided path and prints time spent;}
\usage{
CreateOutputFoldersALL(Year_List, OUTPUT_Path)
}
\arguments{
\item{Year_List}{a list of integers representing the years to be calculated. Folders will be created based; If Windows based, MAXIMUM CHARACTER SIZE IS 118. Or go to the registry and change the maximum number of characters on a file path;}

\item{OUTPUT_Path}{a string representing a path where you'd like the folders to be created;}
}
\value{
has no returns. Meant to be used just to create folders;
}
\description{
Creates some of the folders required to organize outputs in the a certain provided path and prints time spent;
}
\examples{
CreateOutputFoldersALL()
}
