% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportRasterList.R
\name{ExportRasterList}
\alias{ExportRasterList}
\title{Exports a list of rasters and appends the month name to the end, assuming the rasters are ordered by agricultural year; Also prints time spent;}
\usage{
ExportRasterList(
  RasterList,
  Path,
  GeneralFileName,
  StringFinished_One,
  StringFinished_All
)
}
\arguments{
\item{RasterList}{the raster list to be exported as GTiff;}

\item{Path}{a string representing the path to export to;}

\item{GeneralFileName}{a file name for the files, usually xxxx_ as the function will append the month name at the end;}

\item{StringFinished_One}{a string to be printed as each raster is exported;}

\item{StringFinished_All}{a string to be printed in the end;}
}
\value{
doesnt return anything. Meant to be used as an export tool;
}
\description{
Exports a list of rasters and appends the month name to the end, assuming the rasters are ordered by agricultural year; Also prints time spent;
}
