% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateExportPaths.R
\name{GenerateExportPaths}
\alias{GenerateExportPaths}
\title{GenerateExportPaths takes in a list of years and a general OutputPath, that should be the same provided to the function CreateOutputFoldersALL, and automatically defines
in a global environment all the export paths needed for the package to export its several steps.}
\usage{
GenerateExportPaths(Year_List, OutputPath)
}
\arguments{
\item{Year_List}{A list of years.}

\item{OutputPath}{A path to a folder where you would like the exports to go. Should be the same as the one provided to CreateOutputFoldersALL}
}
\value{
This function does not return anything.
}
\description{
GenerateExportPaths takes in a list of years and a general OutputPath, that should be the same provided to the function CreateOutputFoldersALL, and automatically defines
in a global environment all the export paths needed for the package to export its several steps.
}
