% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndividualProducts_1_2_3.R
\name{IndividualZOIProduct3}
\alias{IndividualZOIProduct3}
\title{Takes a list of products, for example Binary_Combo_2016, and cuts it to each individual ZOI, exporting the results to a certain path;}
\usage{
IndividualZOIProduct3(
  ProductList_AllThreshold,
  ZOIShapefileList,
  ExportPath,
  ID_list,
  Threshold_List,
  Threshold_List_2,
  StringName,
  StringOne,
  StringOneZOI,
  StringAllOne,
  StringAllAll
)
}
\arguments{
\item{ProductList_AllThreshold}{a list of rasters representing each month of a particular product, example a list of each month of Binary_Combo_2016;}

\item{ZOIShapefileList}{a list of shapefiles representing the individual ZOI of the original shapefile;}

\item{ExportPath}{a string representing an export path;}

\item{ID_list}{a list of IDs representing the amount of unique shapes inside the original shapefile, usually created as a global variable using CropShapefile_ExportToPath;}

\item{Threshold_List}{a list with a threshold number to be included in the files names, example: 0 as the NDVI threshold to be included the the file name, usually defined earlier in the order of operations, check Vignette;}

\item{Threshold_List_2}{a second list with a threshold number to be included in the files name, example: 800 as the RGBSum threshold, usually defined earlier;}

\item{StringName}{a string to name the files, usually XXX_ as the function will attach the name of the month in the end;}

\item{StringOne}{a string to print as each product is finished;}

\item{StringOneZOI}{a string to print as each complete ZOI is finished;}

\item{StringAllOne}{a string to print as each combination for each ZOI is finished;}

\item{StringAllAll}{a string to print in the end, as every combination for all ZOI are finished;}
}
\value{
a list of lists, with the first list representing ZOI_1 and element 1 of the first list being the first month corresponding to the first ZOI;
}
\description{
Takes a list of products, for example Binary_Combo_2016, and cuts it to each individual ZOI, exporting the results to a certain path;
}
