% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotAllYears_All_Albufeiras_1_2.R
\name{PlotAllYears_All_Albufeiras}
\alias{PlotAllYears_All_Albufeiras}
\title{Plots and exports Water Percentage of the input Dataframes. Produces several plots of a time series of Water Percentage of all the ZOI included in the arguments;}
\usage{
PlotAllYears_All_Albufeiras(Dataframe_List_All, Path)
}
\arguments{
\item{Dataframe_List_All}{a list of lists, with the first element being a list of dataframes. Each dataframe represents a Year of of Product for ZOI 1;}

\item{Path}{a string representing an export path to export the graphs;}
}
\value{
Doesnt return anything, used to plot graphs and export them.
}
\description{
Plots and exports Water Percentage of the input Dataframes. Produces several plots of a time series of Water Percentage of all the ZOI included in the arguments;
}
\examples{
Data too large, check Vignette;
}
