% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSV_to_Dataframe.R
\name{ReadCSVtoDataframe}
\alias{ReadCSVtoDataframe}
\title{Reads a list of CSV files and produce a list of dataframes. CAUTION: You might have to sort the list as it might come unsorted due to file names. Use mixedsort();
CSV_List_Year <<- list.files(path = "XXXX", pattern = "XXXX", full.names = TRUE)
Sorted_CSV_List_Year <<- mixedsort(CSV_List_Year)}
\usage{
ReadCSVtoDataframe(CSV_List)
}
\arguments{
\item{CSV_List}{a list of filepaths to CSV files;}
}
\value{
returns a list of dataframes;
}
\description{
Reads a list of CSV files and produce a list of dataframes. CAUTION: You might have to sort the list as it might come unsorted due to file names. Use mixedsort();
CSV_List_Year <<- list.files(path = "XXXX", pattern = "XXXX", full.names = TRUE)
Sorted_CSV_List_Year <<- mixedsort(CSV_List_Year)
}
\examples{
NULL
}
