% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReclassRasterList.R
\name{ReclassRasterList}
\alias{ReclassRasterList}
\title{Reclassifies a list of rasters acording to the vector provided. The Vectors are generated with the correct formatting by using the function VectorBuilder. Vector1 has NDVI threshold, Vetor 2 has RGMSum threshold and Vetor 3 has the combo to used when reclassifyng the Combo;}
\usage{
ReclassRasterList(RasterList, Vetor, StringFinished_One, StringFinished_All)
}
\arguments{
\item{RasterList}{a raster list to be reclassified acording to vetor provided;}

\item{Vetor}{vetor to orient the reclassification of the raster; Should be presented as follows: Vetor = c(-Inf, ThresholdNumber, 1, ThresholdNumber, Inf, 0);}

\item{StringFinished_One}{a string to print as each raster is reclassified;}

\item{StringFinished_All}{a string to print in the end;}
}
\value{
a list of reclassified rasters acording to the Vetor provided in the arguments;
}
\description{
Reclassifies a list of rasters acording to the vector provided. The Vectors are generated with the correct formatting by using the function VectorBuilder. Vector1 has NDVI threshold, Vetor 2 has RGMSum threshold and Vetor 3 has the combo to used when reclassifyng the Combo;
}
\examples{
Data too large, check Vignette;
}
