% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAlgorithm.R
\name{RunAlgorithm}
\alias{RunAlgorithm}
\title{Function that automates the WaterDetection algorithm. With proper input, everything should be automated. Remenber, NDVI inputs should be named "NDVIListYearXXXX" and Composite inputs
should be named "CompositeListYearXXXX"}
\usage{
RunAlgorithm(
  Year_List,
  ShapefilePath,
  ShapefileName,
  Output_Path,
  NDVI_Threshold,
  RGBSum_Threshold
)
}
\arguments{
\item{Year_List}{a list of years to be calculated;}

\item{ShapefilePath}{path to the original shapefile that contains all the ZOIs to be evaluated;}

\item{ShapefileName}{name of the original shapefile that contains all the ZOIs to be evaluated;}

\item{Output_Path}{general output path where files should be exported;}

\item{NDVI_Threshold}{threshold to be used with NDVI for water detection, recommended zero, 0;}

\item{RGBSum_Threshold}{threshold to be used with the sum of the RGB bands for water detection, recommended eight-hundred, 800;}
}
\value{
this function doesnt return anything, but exports a lot of things.
}
\description{
Function that automates the WaterDetection algorithm. With proper input, everything should be automated. Remenber, NDVI inputs should be named "NDVIListYearXXXX" and Composite inputs
should be named "CompositeListYearXXXX"
}
