% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeRunning.R
\name{TimeRunning}
\alias{TimeRunning}
\title{A gemeral function to collect time spent doing X task. Write StartTime = Sys.time(), then your code, then EndTime = Sys.time(), then}
\usage{
TimeRunning(StartTime, EndTime)
}
\arguments{
\item{StartTime}{A sys.time() time representing starting time;}

\item{EndTime}{A sys.time() time representing ending time}
}
\value{
prints time spent;
}
\description{
A gemeral function to collect time spent doing X task. Write StartTime = Sys.time(), then your code, then EndTime = Sys.time(), then
}
