% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VectorBuilder.R
\name{VectorBuilder}
\alias{VectorBuilder}
\title{Produces global variables Vector1, 2 and 3}
\usage{
VectorBuilder(Threshold_NDVI, Threshold_RGMSum)
}
\arguments{
\item{Threshold_NDVI}{a number representing a threshold to be used to divide NDVI into a binary raster}

\item{Threshold_RGMSum}{a number representing a threshold to be used to divide RGBSum into a binary raster}
}
\value{
Returns nothing, used to produce Vector1, 2 and 3.
}
\description{
Produces global variables Vector1, 2 and 3
}
