% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CropAndMask.R
\name{CropAndMask}
\alias{CropAndMask}
\title{Cuts and Masks a list of rasters to the zone of a shapefile; Cutting the satellite images to the required zones in the beginning saves a LOT of time; Prints time spent;
Might produce an error: "Error in x$.self$finalize() : attempt to apply non-function" but that's an R problem and it doesnt affect the results;}
\usage{
CropAndMask(BrickList, ShapefilePath, ShapeName)
}
\arguments{
\item{BrickList}{a list of Bricks, raster objects, to be reduced, as in cut, to the area of a shapefile;}

\item{ShapefilePath}{a string representing a path to a shapefile;}

\item{ShapeName}{a string representing the shapefiles name;}
}
\value{
a list of rasters cut to the zone of the shapefile presented in the arguments;
}
\description{
Cuts and Masks a list of rasters to the zone of a shapefile; Cutting the satellite images to the required zones in the beginning saves a LOT of time; Prints time spent;
Might produce an error: "Error in x$.self$finalize() : attempt to apply non-function" but that's an R problem and it doesnt affect the results;
}
