% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CropShapefile_ExportToPath.R
\name{CropShapefile_ExportToPath}
\alias{CropShapefile_ExportToPath}
\title{Crops a shapefile into all its individual pieces, by UniqueID, exports all the individual shapefiles into ExportPat plus a new Shapefile with all ZOI and a new column called UniqueID, and prints time spent; Also produces a global variable called ID_List representing a list of IDs taken from the original shapefile;}
\usage{
CropShapefile_ExportToPath(ShapefilePath, ShapefileName, ExportPath)
}
\arguments{
\item{ShapefilePath}{a string representing a path to the shapefile to be cut;}

\item{ShapefileName}{the name of the shapefile file;}

\item{ExportPath}{a string representing the path to which the individual shapefiles are to be exported;}
}
\value{
a list of individual shapefiles, ordered by UniqueID;
}
\description{
Crops a shapefile into all its individual pieces, by UniqueID, exports all the individual shapefiles into ExportPat plus a new Shapefile with all ZOI and a new column called UniqueID, and prints time spent; Also produces a global variable called ID_List representing a list of IDs taken from the original shapefile;
}
